+++ buildSrc/gradle/libs.versions.toml
@@ -6,7 +6,7 @@ junit-api = "org.junit.jupiter:junit-jupiter-api:5.10.3"
 # see https://docs.gradle.org/current/samples/sample_convention_plugins.html#things_to_note.
 pluginPackages-checkerFramework = "org.checkerframework:checkerframework-gradle-plugin:0.6.41"
 pluginPackages-grgit = "org.ajoberstar.grgit:grgit-gradle:5.2.2"
-pluginPackages-jetbrains-changelog = "org.jetbrains.intellij.plugins:gradle-changelog-plugin:2.2.0"
+#pluginPackages-jetbrains-changelog = "org.jetbrains.intellij.plugins:gradle-changelog-plugin:2.2.0"
 #pluginPackages-jetbrains-intellij = "org.jetbrains.intellij.platform:intellij-platform-gradle-plugin:2.0.0-beta6"
 pluginPackages-jetbrains-kotlin = "org.jetbrains.kotlin:kotlin-gradle-plugin:1.9.22"
 pluginPackages-spotless = "com.diffplug.spotless:spotless-plugin-gradle:6.25.0"
