+++ gradle/libs.versions.toml
@@ -28,6 +28,7 @@ vavr = "io.vavr:vavr:0.10.4"
 # where they aren't really `apply`-ed as plugins, but rather added to classpath as plugin packages;
 # see https://docs.gradle.org/current/samples/sample_convention_plugins.html#things_to_note.
 [plugins]
+jetbrains-changelog = "org.jetbrains.changelog:2.2.0"
 jetbrains-grammarkit = "org.jetbrains.grammarkit:2022.3.2.2"
 jetbrains-intellij = "org.jetbrains.intellij.platform:2.0.0-beta6"
 taskTree = "com.dorongold.task-tree:4.0.0"
